Add new icon and update plugin settings
A new icon was added to the project in the form of a SVG file and then utilized within the Constants class. The plugin.xml file was updated to reference this new icon and the tool window ID was adjusted. The plugin icon SVG file was also updated to unify the color gradient across all icons.