Refactor and organize symmetric encryption properties
The commit refactors symmetric encryption properties in the toolset, while improving the ease of addition and management. The properties are now handled as groups, rather than single items, and the ability to add groups has been introduced. It also fixes the plugin compatibility by updating the pluginUntilBuild property to 3000 in gradle.properties file.