Refactor codebase and add JSON Web Token support
The commit involves the refactoring of several classes in the codebase such as LocalConfigFactory, AlgorithmEnum, and others. In addition, a support for JSON Web Tokens (JWT) was added by creating the JWTComponent class and related configuration. Code refactoring also included changes in the data representation and code structure for symmetric cryptography properties and hashing functionality. The new functionalities are reflected in the ToolSetWindow interface. Lastly, a JWTPropDialog class was added for JWT properties manipulation.