Enhanced HTTP component in Toolset UI package with new features
This update introduces significant enhancements to the `HttpComponent` in the UI package of the Toolset codebase. The improvements include added import statements, a number of conditional operations for user settings, an integrated document listener, as well as the development of the sub-class `JT`. This commit further modifies the `TabEnum` and `DataFormatTypeEnum` files in the codebase.