Implement HTTP headers auto-completion feature
This commit includes the essential logic and UI changes to implement HTTP headers auto-completion in HttpComponent. A new CompletionItem interface was created to provide headers for auto-completion. The LanguageTextArea class was modified to include the TextCompletionProvider for registering the auto-completion feature. HTTP headers can now be suggested to the user for ease of use and faster requests setup.