Update Constants class and increment plugin version
The HTTP headers auto-completion feature was implemented in the Constants class. This involves concatenating_HTTP_HEADERS_ values with a colon to facilitate auto-completion. Additionally, the plugin version was updated from 1.1.8 to 1.1.9 in various spots including gradle.properties and plugin.xml file. The changelog was also updated to reflect these changes.