Refactor and add encryption properties model
A significant refactor of the `SymmetricPropDialog`, `JWTPropDialog` and `AESComponent` classes was performed to improve the structure and functionality of the code. A new encryption properties model (`HashCryptoProp`) was added to better handle encryption properties. Other changes include updates to how symmetric encryption properties are read and written, improvements to tree UI components, and better handling of empty states.