Add interval time feature to Remind model and UI
A new attribute, intervalTime, was added to the Remind model and also to the associated user interface, RemindsComponent. This allows the configuration of reminders repeating at regular intervals. Additionally, the scheduleTime was adjusted to calculate intervals more accurately. Necessary changes were also made on the RadioBar component to handle this new feature effectively.